Allow “locale” parameter
PayPal is currently using the wrong language using Safari if the user is not logged in with their PayPal account.

In my case, using Chrome, the PayPal popup shows the Italian language (correct).

Using Safari, the PayPal popup shows the English language (wrong). I have already verified that the language setting on Safari was "Italian".

Using the PayPal "locale" parameter it is at least possible to force the language to be used in the popup.